# backend/openapi/schema.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Podoria CRM API",
        "version": "1.0.0",
        "description": "Versioned API contract for Podoria CRM."
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "operationId": "team_user_list",
                "summary": "List and filter employees for an administrator",
                "parameters": [
                    {
                        "in": "query",
                        "name": "role",
                        "schema": {
                            "enum": [
                                "podologist",
                                "reception",
                                "admin"
                            ],
                            "type": "string",
                            "minLength": 1
                        },
                        "description": "* `podologist` - Подолог\n* `reception` - Рецепція\n* `admin` - Адміністратор"
                    },
                    {
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "in": "query",
                        "name": "status",
                        "schema": {
                            "enum": [
                                "all",
                                "active",
                                "inactive"
                            ],
                            "type": "string",
                            "default": "all",
                            "minLength": 1
                        },
                        "description": "* `all` - all\n* `active` - active\n* `inactive` - inactive"
                    }
                ],
                "tags": [
                    "team"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamUserList"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "team_user_create",
                "summary": "Create an employee with initial password policy",
                "tags": [
                    "team"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TeamUserCreateRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/TeamUserCreateRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TeamUserCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamUser"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/users/{user_id}": {
            "get": {
                "operationId": "team_user_retrieve",
                "summary": "Return one employee profile",
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "team"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamUser"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "patch": {
                "operationId": "team_user_update",
                "summary": "Update contacts, role or active state with last-admin protection",
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "team"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedTeamUserUpdateRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedTeamUserUpdateRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedTeamUserUpdateRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamUser"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/users/{user_id}/deactivate": {
            "post": {
                "operationId": "team_user_deactivate",
                "summary": "Deactivate an employee and revoke their sessions",
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "team"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TeamUser"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorEnvelope": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "fields": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "correlation_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "code",
                    "correlation_id",
                    "fields",
                    "message"
                ]
            },
...
            "PatchedTeamUserUpdateRequest": {
                "type": "object",
                "properties": {
                    "first_name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 150
                    },
                    "last_name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 150
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 32
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1,
                        "maxLength": 254
                    },
                    "role": {
                        "$ref": "#/components/schemas/RoleEnum"
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                }
            },
            "RoleEnum": {
                "enum": [
                    "podologist",
                    "reception",
                    "admin"
                ],
                "type": "string",
                "description": "* `podologist` - Подолог\n* `reception` - Рецепція\n* `admin` - Адміністратор"
            },
...
            "TeamUser": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "$ref": "#/components/schemas/RoleEnum"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "must_change_password": {
                        "type": "boolean"
                    },
                    "temporary_password_expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "last_login": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "required": [
                    "display_name",
                    "email",
                    "first_name",
                    "id",
                    "is_active",
                    "last_login",
                    "last_name",
                    "must_change_password",
                    "phone",
                    "role",
                    "temporary_password_expires_at"
                ]
            },
            "TeamUserCreateRequest": {
                "type": "object",
                "properties": {
                    "first_name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 150
                    },
                    "last_name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 150
                    },
                    "phone": {
                        "type": "string",
                        "default": "",
                        "maxLength": 32
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1,
                        "maxLength": 254
                    },
                    "role": {
                        "$ref": "#/components/schemas/RoleEnum"
                    },
                    "temporary_password": {
                        "type": "string",
                        "writeOnly": true,
                        "minLength": 1
                    },
                    "temporary_password_confirmation": {
                        "type": "string",
                        "writeOnly": true,
                        "minLength": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "default": true
                    },
                    "must_change_password": {
                        "type": "boolean",
                        "default": true
                    }
                },
                "required": [
                    "email",
                    "first_name",
                    "last_name",
                    "role",
                    "temporary_password",
                    "temporary_password_confirmation"
                ]
            },
            "TeamUserList": {
                "type": "object",
                "properties": {
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TeamUser"
                        }
                    }
                },
                "required": [
                    "users"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "podoria_sessionid"
            }
        }
    }
}